Add: Instance variable to track correct guesses
I realized I was using guessed_letters to track
the history of all guesses to display later for
the players but the same variable was also being
used for testing wins.

It made sense to add a new variable to track the
correct guesses and separate the 2 functions into
different variables.