Fix: Bug with update because frozen string literal
The magic comment made it so I couldn't modify the
string in place. I thought the Ruby Style guide
said to only use << and not + for concatenating
strings.

Claude said it turns out << has better performance
and modifies strings in place but what I want is
to create a new string and keeping the magic
comment is a good practice to freeze strings by
default.

The fix was to use + to concatenate and return a
new string.