Fix: Bug where loaded games didn't run
There was a sneaky bug where the player asking to
load a saved game would end up playing a new game.
It turns out the Stackoverflow advice I got to
use slice! was bad. What I was looking for is
the method #delete and not #slice!

Using slice! made the return nil so every time
the if statement checked for a loaded_game name,
it would be nil and the Game class would start a
new game.

Now it's fixed, so it corrects loads the game from
where the saved game left off.